[FIX] 회원가입 시 닉네임 추가 및 닉네임 중복 체크

+++ project_w/src/main/java/shop/winetoy/server/member/entity/RefreshDto.java
@@ -1,14 +1,14 @@
 package shop.winetoy.server.member.entity;
 
 public class RefreshDto {
-	private Long pid;
+	private int pid;
 	private String refreshToken;
 
-	public Long getPid() {
+	public int getPid() {
 		return pid;
 	}
 
-	public void setPid(Long pid) {
+	public void setPid(int pid) {
 		this.pid = pid;
 	}
